Fixes to model loading across the board. Breakage occurred after conservative model loading was fixed to actually be conservative.
git-svn-id: file:///var/www/production/svnmirror/svn_doctrine/branches/0.10/lib@3847 2b2ae520-0f61-44a6-8c59-7e116ffd6265

+++ Doctrine/Migration/Builder.php
@@ -155,8 +155,8 @@ END;
      */
     public function generateMigrationsFromModels($modelsPath = null)
     {
-        if ($modelsPath) {
-            $models = Doctrine::loadModels($modelsPath);
+        if ($modelsPath !== null) {
+            $models = Doctrine::filterInvalidModels(Doctrine::loadModels($modelsPath));
         } else {
             $models = Doctrine::getLoadedModels();
         }
@@ -172,26 +172,28 @@ END;
             $down = $this->buildDropTable($export);
             
             $className = 'Add' . Doctrine::classify($export['tableName']);
-            
+
             $this->generateMigrationClass($className, array(), $up, $down);
         }
         
-        $className = 'ApplyForeignKeyConstraints';
+        if ( ! empty($foreignKeys)) {
+            $className = 'ApplyForeignKeyConstraints';
         
-        $up = '';
-        $down = '';
-        foreach ($foreignKeys as $tableName => $definitions)    {
-            $tableForeignKeyNames[$tableName] = array();
+            $up = '';
+            $down = '';
+            foreach ($foreignKeys as $tableName => $definitions)    {
+                $tableForeignKeyNames[$tableName] = array();
             
-            foreach ($definitions as $definition) {
-                $definition['name'] = $tableName . '_' . $definition['foreignTable'] . '_' . $definition['local'] . '_' . $definition['foreign'];
+                foreach ($definitions as $definition) {
+                    $definition['name'] = $tableName . '_' . $definition['foreignTable'] . '_' . $definition['local'] . '_' . $definition['foreign'];
                 
-                $up .= $this->buildCreateForeignKey($tableName, $definition);
-                $down .= $this->buildDropForeignKey($tableName, $definition);
+                    $up .= $this->buildCreateForeignKey($tableName, $definition);
+                    $down .= $this->buildDropForeignKey($tableName, $definition);
+                }
             }
-        }
         
-        $this->generateMigrationClass($className, array(), $up, $down);
+            $this->generateMigrationClass($className, array(), $up, $down);
+        }
         
         return true;
     }
@@ -272,6 +274,10 @@ END;
             $class = $this->buildMigrationClass($className, $fileName, $options, $up, $down);
             
             $path = $this->getMigrationsPath() . DIRECTORY_SEPARATOR . $fileName;
+            
+            if ( class_exists($className)) {
+                throw new Doctrine_Exception('Migration class with the name "' . $className . '" already exists.');
+            }
             
             file_put_contents($path, $class);
         }
